requirements.txt
fastapi
uvicorn
google-generativeai
mysql-connector-python
python-dotenv
pydantic
